[#9] refactor: service -> application 패키지 네이밍 수정
